requirements.txt
scikit-learn==1.4.2
pandas==2.2.2
mlflow==2.14.1
matplotlib==3.9.0
pyyaml==6.0.2
pytest==8.3.2
ruff==0.5.0
